Create ArmstrongNumber
Armstrong Number
An Armstrong Number (also known as a Narcissistic Number or Pluperfect Number) is a number that is equal to the sum of its own digits each raised to the power of the number of digits. In other words, for a given number of n digits, an Armstrong number satisfies the condition:

Armstrong Number
=
(𝑑1)𝑛+(𝑑2)𝑛+...+(𝑑𝑛)𝑛
Armstrong Number=(d 1) n+(d 2) n+...+(d n) n
 
153 is an Armstrong number because it has 3 digits, and:
153
=
(
1
)
3
+
(
5
)
3
+
(
3
)
3
=
1
+
125
+
27
=
153
153=(1) 
3
 +(5) 
3
 +(3) 
3
 =1+125+27=153
9474 is an Armstrong number because it has 4 digits, and:
9474
=
(
9
)
4
+
(
4
)
4
+
(
7
)
4
+
(
4
)
4
=
6561
+
256
+
2401
+
256
=
9474
9474=(9) 
4
 +(4) 
4
 +(7) 
4
 +(4) 
4
 =6561+256+2401+256=9474
Properties:

Armstrong numbers are used primarily for educational purposes to understand loops, conditional statements, and mathematical computations in programming.
They exist only for certain ranges, and the count of such numbers diminishes as the number of digits increases.