Create FactorofaNumber
Factors of a number are the integers that divide the number exactly without leaving a remainder. In other words, if x is a factor of n, then n % x == 0. Factors are used to determine divisibility, simplify fractions, and solve various mathematical problems.

How to Find Factors
To find the factors of a number n:

Iterate through all integers from 1 to n.
Check if the number n is divisible by the integer (i.e., n % i == 0).
If divisible, then the integer is a factor of n.